requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# LLM providers
openai>=1.10.0

# Vector store
faiss-cpu>=1.7.4
tiktoken>=0.5.2

# CLI interface
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: Only needed if using Ollama
# ollama>=0.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0